refactor(request-callout): extract plugin wrapper as a shared component
